+++ py7zr/py7zr.py
@@ -2,7 +2,7 @@
 #
 # p7zr library
 #
-# Copyright (c) 2019-2022 Hiroshi Miura <miurahr@linux.com>
+# Copyright (c) 2019-2024 Hiroshi Miura <miurahr@linux.com>
 # Copyright (c) 2004-2015 by Joachim Bauch, mail@joachim-bauch.de
 # 7-Zip Copyright (C) 1999-2010 Igor Pavlov
 # LZMA SDK Copyright (C) 1999-2010 Igor Pavlov
@@ -57,8 +57,6 @@ from py7zr.exceptions import (
 )
 from py7zr.helpers import (
     ArchiveTimestamp,
-    MemIO,
-    NullIO,
     calculate_crc32,
     check_archive_path,
     filetime_to_dt,
@@ -67,6 +65,7 @@ from py7zr.helpers import (
     readlink,
     remove_trailing_slash,
 )
+from py7zr.io import MemIO, NullIO, WriterFactory
 from py7zr.properties import DEFAULT_FILTERS, FILTER_DEFLATE64, MAGIC_7Z, get_default_blocksize, get_memory_limit
 
 if sys.platform.startswith("win"):
@@ -74,7 +73,6 @@ if sys.platform.startswith("win"):
 
 FILE_ATTRIBUTE_UNIX_EXTENSION = 0x8000
 FILE_ATTRIBUTE_WINDOWS_MASK = 0x07FFF
-
 
 class ArchiveFile:
     """Represent each files metadata inside archive file.
@@ -393,7 +391,6 @@ class SevenZipFile(contextlib.AbstractContextManager):
             raise TypeError(f"invalid file: {type(file)}")
         self.encoded_header_mode = True
         self.header_encryption = header_encryption
-        self._fileRefCnt = 1
         try:
             if mode == "r":
                 self._real_get_contents(password)
@@ -416,7 +413,6 @@ class SevenZipFile(contextlib.AbstractContextManager):
         except Exception as e:
             self._fpclose()
             raise e
-        self._dict: dict[str, IO[Any]] = {}
         self.dereference = dereference
         self.reporterd: Optional[Thread] = None
         self.q: queue.Queue[Any] = queue.Queue()
@@ -428,9 +424,7 @@ class SevenZipFile(contextlib.AbstractContextManager):
         self.close()
 
     def _fpclose(self) -> None:
-        assert self._fileRefCnt > 0
-        self._fileRefCnt -= 1
-        if not self._fileRefCnt and not self._filePassed:
+        if not self._filePassed:
             self.fp.close()
 
     def _real_get_contents(self, password) -> None:
@@ -536,11 +530,11 @@ class SevenZipFile(contextlib.AbstractContextManager):
         self,
         path: Optional[Any] = None,
         targets: Optional[Collection[str]] = None,
-        return_dict: bool = False,
         callback: Optional[ExtractCallback] = None,
         enable_symlink: bool = False,
         recursive: Optional[bool] = False,
-    ) -> Optional[dict[str, IO[Any]]]:
+        writer_factory: Optional[WriterFactory] = None,
+    ) -> None:
         if callback is None:
             pass
         elif isinstance(callback, ExtractCallback):
@@ -592,15 +586,13 @@ class SevenZipFile(contextlib.AbstractContextManager):
                 outfilename = get_sanitized_output_path(outname, path)
             else:
                 outfilename = get_sanitized_output_path(outname, pathlib.Path(os.getcwd()).joinpath(path))
-            if return_dict:
+            if writer_factory is not None:
                 if f.is_directory or f.is_socket:
                     # ignore special files and directories
                     pass
                 else:
                     fname = outfilename.as_posix()
-                    _buf = io.BytesIO()
-                    self._dict[fname] = _buf
-                    self.worker.register_filelike(f.id, MemIO(_buf))
+                    self.worker.register_filelike(f.id, MemIO(fname, writer_factory))
             elif f.is_directory:
                 if not outfilename.exists():
                     target_dirs.append(outfilename)
@@ -646,8 +638,8 @@ class SevenZipFile(contextlib.AbstractContextManager):
 
         self.q.put(("post", None, None))
         # early return when dict specified
-        if return_dict:
-            return self._dict
+        if writer_factory is not None:
+            return
         # set file properties
         for outfilename, properties in target_files:
             # mtime
@@ -1022,40 +1014,38 @@ class SevenZipFile(contextlib.AbstractContextManager):
             )
         return alist
 
-    def readall(self) -> Optional[dict[str, IO[Any]]]:
-        self._dict = {}
-        return self._extract(path=None, return_dict=True)
-
     def extractall(
-        self, path: Optional[Any] = None, callback: Optional[ExtractCallback] = None, enable_symlink: bool = True
+        self,
+        path: Optional[Any] = None,
+        *,
+        callback: Optional[ExtractCallback] = None,
+        factory: Optional[WriterFactory] = None,
+        enable_symlink: bool = True
     ) -> None:
         """Extract all members from the archive to the current working
         directory and set owner, modification time and permissions on
-        directories afterwards. ``path`` specifies a different directory
+        directories afterward. ``path`` specifies a different directory
         to extract to.
         """
-        self._extract(path=path, return_dict=False, callback=callback, enable_symlink=enable_symlink)
-
-    def read(self, targets: Optional[Collection[str]] = None) -> Optional[dict[str, IO[Any]]]:
-        if not self._is_none_or_collection(targets):
-            raise TypeError("Wrong argument type given.")
-        # For interoperability with ZipFile, we strip any trailing slashes
-        # This also matches the behavior of TarFile
-        if targets is not None:
-            targets = [remove_trailing_slash(target) for target in targets]
-        self._dict = {}
-        return self._extract(path=None, targets=targets, return_dict=True)
+        self._extract(path=path, callback=callback, writer_factory=factory, enable_symlink=enable_symlink)
 
     def extract(
-        self, path: Optional[Any] = None, targets: Optional[Collection[str]] = None, recursive: Optional[bool] = False,
-      enable_symlink: bool = True) -> None:
+        self,
+        path: Optional[Any] = None,
+        targets: Optional[Collection[str]] = None,
+        recursive: Optional[bool] = False,
+        *,
+        callback: Optional[ExtractCallback] = None,
+        factory: Optional[WriterFactory] = None,
+        enable_symlink: bool = True,
+    ) -> None:
         if not self._is_none_or_collection(targets):
             raise TypeError("Wrong argument type given.")
         # For interoperability with ZipFile, we strip any trailing slashes
         # This also matches the behavior of TarFile
         if targets is not None:
             targets = [remove_trailing_slash(target) for target in targets]
-        self._extract(path, targets, return_dict=False, recursive=recursive, enable_symlink=enable_symlink)
+        self._extract(path, targets, recursive=recursive, callback=callback, writer_factory=factory, enable_symlink=enable_symlink)
 
     def reporter(self, callback: ExtractCallback):
         while True:
@@ -1112,10 +1102,6 @@ class SevenZipFile(contextlib.AbstractContextManager):
         self.files.append(file_info)
         self.worker.archive(self.fp, self.files, folder, deref=dereference)
 
-    def writed(self, targets: dict[str, IO[Any]]) -> None:
-        for target, input in targets.items():
-            self.writef(input, target)
-
     def writef(self, bio: IO[Any], arcname: str):
         if not check_archive_path(arcname):
             raise ValueError(f"Specified path is bad: {arcname}")
@@ -1135,6 +1121,8 @@ class SevenZipFile(contextlib.AbstractContextManager):
             last = bio.tell()
             bio.seek(current, os.SEEK_SET)
             size = last - current
+        elif isinstance(bio, MemIO):
+            size = bio.__sizeof__()
         else:
             raise ValueError("Wrong argument passed for argument bio.")
         if size >= 0:
@@ -1420,7 +1408,8 @@ class Worker:
                 # delayed execution of crc check.
                 self._check(fp, just_check, src_end)
                 just_check = []
-                fileish.parent.mkdir(parents=True, exist_ok=True)
+                if not isinstance(fileish, MemIO):
+                    fileish.parent.mkdir(parents=True, exist_ok=True)
                 if not f.emptystream:
                     if f.is_junction and not isinstance(fileish, MemIO) and sys.platform == "win32":
                         with io.BytesIO() as ofp: